FEATURE: Add global and worker timeout option.
This PR addresses the feedback what was originally given on PR #162 by
adding a default worker specific timeout option in addition to a global
timeout.

The original commit

https://github.com/discourse/image_optim/commit/8bf3c0e32f50cd2e3e23c0b5b936947994ff75e8

on PR #162 with just the global timeout has been in production since
2018-07-08 on Discourse instances using a forked version of this gem.

We would like to get this change merged in, so that we can get off of
the discourse specific fork and use the latest version of the
image_optim gem.

Co-authored-by: Blake Erickson <o.blakeerickson@gmail.com>

Do not raise if optimized before timeout

Based on feedback this commit includes the following changes:

When optimizing multiple images with `optimize_images` a single image
timeout doesn't break execution. The global timeout will still raise an
exception though if it is reached.

When optimizing a single image do not raise an exception if a single
worker has succeeded but has reached the timeout period while running
through the remaining workers.

When optimizing a single image if a single worker times out it will
continue on to the next worker until the global timeout is reached.